+++ noodles-vcf/CHANGELOG.md
@@ -4,7 +4,10 @@
 
 ### Added
 
-  * vcf/record/position: Implement `PartialEq<core::Position>` for `Position`.
+  * vcf/record/position: Implement `PartialEq<core::Position>` and
+    `PartialOrd<core::Position>` for `Position` ([#191]).
+
+[#191]: https://github.com/zaeleus/noodles/issues/191
 
 ## 0.35.0 - 2023-08-03